refactor all service from bunch of files to single files
